# api/api-codegen/api/swagger.yaml
openapi: 3.0.0
info:
  title: Simple Twitter API
  description: API for a simple Twitter-like application.
  version: 1.0.0
servers:
- url: https://api.simpletwitter.com/v1
paths:
  /users:
    post:
      summary: Create a new user
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/users_body'
        required: true
      responses:
        "201":
          description: Created
  /users/{user_id}:
    get:
      summary: Get a user by ID
      operationId: getUser
      parameters:
      - name: user_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /tweets:
    post:
      summary: Create a new tweet
      operationId: createTweet
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/tweets_body'
        required: true
      responses:
        "201":
          description: Created
  /tweets/{tweet_id}:
    get:
      summary: Get a tweet by ID
      operationId: getTweet
      parameters:
      - name: tweet_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tweet'
  /follows:
    get:
      tags:
      - Follows
      summary: Get all followers for a specific user
      operationId: getFollowsForUser
      parameters:
      - name: user_id
        in: query
        description: The ID of the user to retrieve followers for
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: A list of followers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Follow'
                x-content-type: application/json
    post:
      summary: Create a follow relationship
      operationId: createFollow
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/follows_body'
        required: true
      responses:
        "201":
          description: Created
  /followers:
    get:
      tags:
      - Followers
      summary: Get all followees for a specific user
      operationId: getFollowersForUser
      parameters:
      - name: user_id
        in: query
        description: The ID of the user to retrieve followees for
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: A list of followees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Follower'
                x-content-type: application/json
components:
  schemas:
    User:
      type: object
      properties:
        user_id:
          type: string
        user_name:
          type: string
        email:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      example:
        updated_at: 2000-01-23T04:56:07.000+00:00
        user_id: user_id
        user_name: user_name
        created_at: 2000-01-23T04:56:07.000+00:00
        email: email
    Tweet:
      type: object
      properties:
        tweet_id:
          type: string
        user_id:
          type: string
        tweet_text:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      example:
        updated_at: 2000-01-23T04:56:07.000+00:00
        tweet_id: tweet_id
        user_id: user_id
        created_at: 2000-01-23T04:56:07.000+00:00
        tweet_text: tweet_text
    Follow:
      type: object
      properties:
        user_id:
          type: string
        follow_id:
          type: string
      example:
        user_id: user_id
        follow_id: follow_id
    Follower:
      type: object
      properties:
        user_id:
          type: string
        follower_id:
          type: string
      example:
        follower_id: follower_id
        user_id: user_id
    users_body:
      required:
      - email
      - user_name
      type: object
      properties:
        user_name:
          type: string
        email:
          type: string
    tweets_body:
      required:
      - tweet_text
      - user_id
      type: object
      properties:
        user_id:
          type: string
        tweet_text:
          type: string
    follows_body:
      required:
      - follow_id
      - user_id
      type: object
      properties:
        user_id:
          type: string
        follow_id:
          type: string