# api/api-codegen/go/model_tweet.go
/*
 * Simple Twitter API
 *
 * API for a simple Twitter-like application.
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package cmd
import (
	"time"
)

type Tweet struct {

	TweetId string `json:"tweet_id,omitempty"`

	UserId string `json:"user_id,omitempty"`

	TweetText string `json:"tweet_text,omitempty"`

	CreatedAt time.Time `json:"created_at,omitempty"`

	UpdatedAt time.Time `json:"updated_at,omitempty"`
}