# api/api-codegen/go/model_user.go
/*
 * Simple Twitter API
 *
 * API for a simple Twitter-like application.
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package cmd
import (
	"time"
)

type User struct {

	UserId string `json:"user_id,omitempty"`

	UserName string `json:"user_name,omitempty"`

	Email string `json:"email,omitempty"`

	CreatedAt time.Time `json:"created_at,omitempty"`

	UpdatedAt time.Time `json:"updated_at,omitempty"`
}